finish tidy-up of test_features
